requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2024.10
Django==4.2.20
django-crispy-forms==2.3
psycopg2==2.9.10
python-dotenv==1.0.1
pytz==2025.1
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
